requirements.txt
numpy
pandas
matplotlib
seaborn
pymongo
nltk
torch
datasets
evaluate
transformers
scikit-learn
python-dotenv
#-e .